+++ src/types/Interactive.ts
@@ -1,0 +1,23 @@
+export type Interactive = {
+    type: 'list' | 'button' | 'product' | 'flow' | 'location_request_message',
+    body?: {
+        text: string
+    },
+    action: {
+        name?: string,
+        buttons?: Array<{
+            type: 'reply',
+            reply?: {
+                id: string,
+                title: string
+            }
+        }>
+    },
+    header?: {
+        type: 'text' | 'image' | 'video' | 'document',
+        text: string
+    }
+    footer?: {
+        text: string
+    }
+}