+++ src/views/settings/messages/Edit.vue
@@ -1,59 +1,50 @@
 <template>
   <!-- Modal -->
   <div :id="selectedMessage.id" class="modal fade" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-     <div class="modal-dialog modal-lg">
+     <div class="modal-dialog modal-xl">
        <div class="modal-content">
          <div class="modal-header">
            <h5 class="modal-title">{{ $t('common.titles.title_modal') }}</h5>
            <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
          </div>
-         <div class="modal-body row">
-           <div class="col-md-6">
+         <div class="modal-body">
+          <div class="row">
+            <div class="col-md-6">
              <label for="message-name" class="col-form-label">{{ $t('common.fields.name') }}</label>
-             <input ref="textArea" name="message-name" class="form-control"
+             <input ref="inputName" name="message-name" class="form-control"
                     v-model="$props.selectedMessage.name"/>
-             <label for="message-text" class="col-form-label">{{ $t('common.fields.label_message') }}</label>
-             <textarea rows="5" ref="textArea" id="editorText" class="form-control text-area-Message" contenteditable="true"  @input="updateTextareaMessage"
-              v-model="newMessage"/>
-             <div class="d-flex mt-1">
-               <button class="bold-button btn btn-sm btn-info btn-squared px-4 py-2 active"
-                       @click="letterBold">
-                 <b>B</b>
-               </button>
-               <button class="italic-button btn btn-sm btn-info btn-squared px-4 py-2 active ms-1"
-                       @click="letterItalic">
-                       <span class="fs-6">I</span><i></i>
-               </button>
-               <button class="emoji-button btn btn-sm btn-info btn-squared px-4 py-2 ms-1" @click="toggleEmojiPicker">
-                😀
-               </button>
-             </div>
-             <div v-show="isEmojiPickerOpen" class="emoji-picker position-absolute">
-               <EmojiPicker @select="insertEmoji" />
-             </div>
-             <div class="d-flex align-items-center flex-wrap">
-               <div v-for="(placeholder, index) in placeholders" :key="index">
-                 <span class="tooltip-element badge bg-secondary ms-1" data-bs-toggle="tooltip" :title="$t(placeholder.description)" @click="insertPlaceholder(placeholder.value)">
-                   {{ $t(placeholder.label) }}
-                 </span>
-               </div>
-             </div>
+            <TextEditor :selectedMessage="$props.selectedMessage.message" @messageUpdated="updateMessage" @contentUpdated="updateFormattedMessage" />
            </div>
            <div class="col-md-6">
              <label for="message-text" class="col-form-label">{{ $t('common.fields.label_preview') }}</label>
-             <div class="preview-container form-control" v-html="formattedMessage" disabled></div>
+             <div class="whatsapp-container">
+              <div class="whatsapp-message">
+                <div class="preview-container" v-html="formattedMessage" disabled></div>
+              </div>
+            </div>
              <div class="mb-3" v-if="$props.selectedMessage">
                <label for="description-text" class="col-form-label">{{ $t('common.fields.label_description') }}</label>
                <textarea  class="form-control text-area-Description" id="description-text" aria-label="Description"
-                         aria-describedby="description-addon" v-model="$props.selectedMessage.description" rows="5" @input="updateTextareaMessage" />
+                         aria-describedby="description-addon" v-model="$props.selectedMessage.description" rows="5" @input="updateTextareaDescription" />
              </div>
            </div>
+          </div>
+          <hr>
+          <div class="row">
+            <div class="col-md-12">
+              <button type="button" class="btn btn-outline-primary" @click="toggleInteractiveMessage">
+                {{ isInteractiveMessage ? $t('wp.actions.disable_interactive') : $t('wp.actions.enable_interactive') }}
+              </button>
+            </div>
+            <InteractiveMessageBuilder @interactive-updated="updateInteractiveMessage" :selectedInteractive="$props.selectedMessage.interactive" v-if="isInteractiveMessage"/>
+          </div>
+
          </div>
          <div class="modal-footer">
            <button type="button" class="btn bg-gradient-secondary" data-bs-dismiss="modal">
              {{ $t('common.actions.close') }}
            </button>
-           <button type="submit" class="btn bg-gradient-primary" @click="saveChanges">{{ $t('common.actions.submit') }}</button>
+           <button type="button" class="btn bg-gradient-primary" @click="saveChanges">{{ $t('common.actions.submit') }}</button>
          </div>
        </div>
      </div>
@@ -61,36 +52,39 @@
  </template>
 
 <script setup lang="ts">
-import {ref, computed, defineProps, defineEmits, Ref, onMounted} from 'vue'
+import {ref, defineProps, defineEmits, Ref, onMounted, watch} from 'vue'
 import SettingsRepository from '@/repositories/SettingsRepository'
 import ToastService from '@/services/ToastService'
 import { useLoadingState } from '@/services/stores/LoadingState'
 import { SettingsMessageInterface } from '@/types/SettingsMessagesInterface'
 import { hide } from '@/helpers/ModalHelper'
 import i18n from '@/plugins/i18n'
-import EmojiPicker, {EmojiExt} from 'vue3-emoji-picker'
+import InteractiveMessageBuilder from '@/components/InteractiveMessageBuilder.vue'
+import TextEditor from '@/components/TextEditor.vue'
+import { Interactive } from '@/types/Interactive'
 
+const props = defineProps<{ selectedMessage: SettingsMessageInterface}>()
 const { setLoading } = useLoadingState()
 const textArea = ref<HTMLTextAreaElement | null>(null)
-const isEmojiPickerOpen = ref(false)
+const text = ref<string>('')
+const formattedMessage = ref<string>('')
 const emit = defineEmits(['updateMessages'])
-const props = defineProps<{ selectedMessage: SettingsMessageInterface}>()
-const newMessage: Ref<string> = ref('')
+const isInteractiveMessage = ref<boolean>(false)
+const interactiveMessage = ref<Interactive|null>(null)
 
-const placeholders = [
-  { description: 'common.placeholders.description.plate', label: 'common.placeholders.label.plate', value: '[[PLATE]]' },
-  { description: 'common.placeholders.description.vehicle_color', label: 'common.placeholders.label.color', value: '[[COLOR]]' },
-  { description: 'common.placeholders.description.username', label: 'common.placeholders.label.name', value: '[[USERNAME]]' },
-  { description: 'common.placeholders.description.company_number', label: 'common.placeholders.label.number_pqr', value: '[[PQR-NUMBER]]' },
-  { description: 'common.placeholders.description.place_name', label: 'common.placeholders.label.place_name', value: '[[PLACE]]' },
-  { description: 'common.placeholders.description.company_name', label: 'common.placeholders.label.company_name', value: '[[COMPANY]]' },
-]
+function updateFormattedMessage(text: string): void{
+  formattedMessage.value = text
+}
 
-onMounted(async() => {
-  newMessage.value = props.selectedMessage.message?? ''
-})
+function updateMessage(content: string): void {
+  text.value = content
+}
 
-function updateTextareaMessage() {
+function updateInteractiveMessage(message: Interactive|null): void {
+  interactiveMessage.value = message
+}
+
+function updateTextareaDescription() {
   const textarea = textArea.value
   if (textarea) {
     textarea.style.height = 'auto'
@@ -98,103 +92,46 @@ function updateTextareaMessage() {
   }
 }
 
-function insertPlaceholder(placeholder: string): void {
-  if (placeholder && textArea.value) {
-    const startPos = textArea.value.selectionStart ?? 0
-    const endPos = textArea.value.selectionEnd ?? 0
-    const scrollTop = textArea.value.scrollTop
-    const value = textArea.value.value
-    textArea.value.value = value.substring(0, startPos) + placeholder + value.substring(endPos, value.length)
-    textArea.value.focus()
-    textArea.value.setSelectionRange(startPos + placeholder.length, startPos + placeholder.length)
-    textArea.value.scrollTop = scrollTop
+function toggleInteractiveMessage(): void {
+  isInteractiveMessage.value = !isInteractiveMessage.value
+  if (!isInteractiveMessage.value) {
+    interactiveMessage.value = null
+  } else {
+    interactiveMessage.value = {
+      type: 'button',
+      body: { text: text.value },
+      action: {
+        buttons: [],
+      },
+    }
   }
-}
-
-function applyStyleToSelection(style: string, text?: string): void {
-  if (!textArea.value) return
-  const startPos = textArea.value.selectionStart ?? 0
-  const endPos = textArea.value.selectionEnd ?? 0
-  const selectedText = text || (textArea.value.value.substring(startPos, endPos))
-  const styledText = style === 'b' ? `*${selectedText}*` : `_${selectedText}_`
-  const value = textArea.value.value
-  textArea.value.value = value.substring(0, startPos) + styledText + value.substring(endPos, value.length)
-  textArea.value.focus()
-  const newEndPos = startPos + styledText.length - 1
-  textArea.value.setSelectionRange(newEndPos, newEndPos)
-}
-
-function applyStyle(style: string): void {
-  if (style === 'b' || style === 'i') {
-    applyStyleToSelection(style)
-  }
-}
-
-function insertTextAtCursor(text: string): void {
-  newMessage.value += text
-  if (textArea.value) {
-    textArea.value.focus()
-  }
-}
-
-function letterBold(): void {
-  applyStyle('b')
-}
-
-function letterItalic(): void {
-  applyStyle('i')
-}
-
-function insertEmoji(emoji: EmojiExt): void {
-  if (emoji.i) {
-    insertTextAtCursor(emoji.i)
-    toggleEmojiPicker()
-  }
-}
-
-function toggleEmojiPicker(): void {
-  isEmojiPickerOpen.value = !isEmojiPickerOpen.value;
 }
 
 function saveChanges(): void {
-  if (props.selectedMessage && textArea.value) {
-    setLoading(true)
-    const updatedMessage = {
-      id: props.selectedMessage.id,
-      name: props.selectedMessage.name,
-      description: props.selectedMessage.description,
-      message: textArea.value.value,
-      enabled: props.selectedMessage.enabled 
-    }
-    SettingsRepository.updateMessage(updatedMessage).then(async () => {
-      emit('updateMessages')
-      setLoading(false)
-      hide(updatedMessage.id)
-      await ToastService.toast(ToastService.SUCCESS, i18n.global.t('common.messages.updated'))
-    }).catch(async e => {
-      setLoading(false)
-      await ToastService.toast(ToastService.ERROR, i18n.global.t('common.messages.error'), e.message)
-    })
+  if (!props.selectedMessage.message && text.value == '') return
+  setLoading(true)
+  const updatedMessage = {
+    id: props.selectedMessage.id,
+    name: props.selectedMessage.name,
+    description: props.selectedMessage.description,
+    message: text.value,
+    enabled: props.selectedMessage.enabled,
+    interactive: interactiveMessage.value,
   }
+
+  SettingsRepository.updateMessage(updatedMessage).then(async () => {
+    emit('updateMessages')
+    setLoading(false)
+    hide(updatedMessage.id)
+    await ToastService.toast(ToastService.SUCCESS, i18n.global.t('common.messages.updated'))
+  }).catch(async e => {
+    setLoading(false)
+    await ToastService.toast(ToastService.ERROR, i18n.global.t('common.messages.error'), e.message)
+  })
 }
 
-const formattedMessage = computed(() => {
-  let content = newMessage.value
-  const tabs = content.match(/\t/g) || []
-  const newLines = content.match(/\n/g) || []
-
-  content = content
-  .replace(/\*([^*]+)\*/g, '<b>$1</b>')
-  .replace(/_([^_]+)_/g, '<i>$1</i>')
-
-  for (let i = 0; i < tabs.length; i++) {
-    content = content.replace(/\t/, '&#9;')
-  }
-
-  for (let i = 0; i < newLines.length; i++) {
-    content = content.replace(/\n/, '<br>')
-  }
-
-  return content;
+onMounted(() => {
+  isInteractiveMessage.value = props.selectedMessage.interactive != undefined
+  text.value = props.selectedMessage.message    
 })
 </script>